refactor: #55 Vision/Mac 기능 구조 재구성 및 Immersive Scene 런타임 통합   - Features/MacUI → Features/Mac/UI로 재구성   - Features/VisionUI → Features/Vision/UI로 재구성   - Vision Runtime 레이어 추가 (AppRouter, ImmersiveCoordinator, RealityKit 통합)   - ImmersiveView 및 RootView 신규 추가   - DI(Dependency Injection) 계층 구현 (VisionEnvironmentKey, View+VisionDependencies)   - ImmersiveIDs 공통 상수 레이어 추가
