+++ Hippo/Applications/VisionApp/HippoVisionApp.swift
@@ -13,10 +13,14 @@ struct HippoVisionApp: App {
 
     var body: some Scene {
         WindowGroup(id: "InitialWindow") {
-            HomeView()
+            RootView()
                 .frame(minWidth: 580, maxWidth: 1020, minHeight: 760, maxHeight: 1020)
                 .environment(appModel)
         }
         .windowResizability(.contentSize)
+        
+        ImmersiveSpace(id: ImmersiveIDs.surgery) {
+            ImmersiveView()
+        }
     }
 }
