+++ Hippo/Features/Vision/Runtime/ImmersiveSceneRuntime.swift
@@ -1,0 +1,30 @@
+import Combine
+import RealityKit
+import SwiftUI
+
+@MainActor
+final class ImmersiveSceneRuntime: ObservableObject {
+
+    // RealityView 의 content 관리
+    func setupScene(in content: RealityViewContent) {
+        // RealityViewContent 는 RealityView 의 클로저 파라미터입니다.
+        // 여기서 scene을 직접 new 하는 게 아니라, AnchorEntity를 추가해야 해요.
+        let anchor = AnchorEntity(world: .zero)
+
+        // 테스트용 엔티티
+        let sphere = ModelEntity(mesh: .generateSphere(radius: 0.1))
+        sphere.position = [0, 1, -1]
+        anchor.addChild(sphere)
+
+        // RealityViewContent에 엔티티 추가
+        content.add(anchor)
+    }
+
+    func start() {
+        print("ImmersiveSceneRuntime started")
+    }
+
+    func stop() {
+        print("ImmersiveSceneRuntime stopped")
+    }
+}
