+++ Hippo/Features/Vision/UI/Home/Components/Buttons/AddPatientButton.swift
