+++ Hippo/Features/Vision/UI/Home/Components/Headers/LogoHeader.swift
