+++ Hippo/Features/Vision/UI/Home/Components/Headers/SectionHeader.swift
