+++ Hippo/Features/Vision/UI/Home/Components/PatientSection/OperationDateBadge.swift
