+++ Hippo/Features/Vision/UI/Home/Components/PatientSection/PatientCell.swift
