+++ Hippo/Features/Vision/UI/Home/Components/PatientSection/PatientSectionView.swift
