+++ Hippo/Shared/Presentation/Constants/ImmersiveIDs.swift
@@ -1,0 +1,9 @@
+//
+//  ImmersiveIDs.swift
+//  Hippo
+//
+//  Created by eunsong on 10/22/25.
+//
+enum ImmersiveIDs {
+    static let surgery = "SurgeryImmersiveSpace"
+}
