+++ HippoTests/TeleVisionTests.swift
@@ -5,13 +5,13 @@
 //  Created by 김현기 on 10/13/25.
 //
 
-import Testing
-@testable import TeleVision
-
-struct TeleVisionTests {
-
-    @Test func example() async throws {
-        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    }
-
-}
+//import Testing
+//@testable import TeleVision
+//
+//struct TeleVisionTests {
+//
+//    @Test func example() async throws {
+//        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+//    }
+//
+//}
